to template

+++ src/main/java/com/edu/mesler/currency/domain/Currency.java
@@ -3,13 +3,12 @@ package com.edu.mesler.currency.domain;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 public class Currency {
-    Long id;
+    int id;
     String code;
     String fullName;
     String sign;
